% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{axis}
\alias{axis}
\alias{eyAxis_category}
\alias{eyAxis_time}
\alias{eyAxis_value}
\title{Customise axis}
\usage{
eyAxis_category(p, data, show = TRUE, zlevel = 0, z = 0,
  boundaryGap = FALSE, ...)

eyAxis_value(p, data, show = TRUE, zlevel = 0, z = 0, position = "left",
  name = NULL, nameLocation = "end", nameTextStyle = list(),
  boundaryGap = list(0, 0), min = NULL, max = NULL, scale = FALSE,
  splitNumber = NULL, ...)

eyAxis_time(p, data, show = TRUE, zlevel = 0, z = 0, position = "left",
  name = NULL, nameLocation = "end", nameTextStyle = list(),
  boundaryGap = list(0, 0), min = NULL, max = NULL, scale = FALSE,
  splitNumber = NULL, ...)
}
\description{
Customise axis

Y axis categorical

Y axis numerical

Y axis date and time
}

