% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{eanimation}
\alias{eanimation}
\title{Add animations}
\usage{
eanimation(p, animation = TRUE, addDataAnimation = TRUE,
  animationThreshold = 2000, animationDuration = 2000,
  animationDurationUpdate = 500, animationEasing = "ExponentialOut", ...)
}
\description{
Add animations
}

