% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{earea}
\alias{earea}
\alias{earea_}
\title{Add area}
\usage{
earea(p, serie, name = NULL, stack = NULL, smooth = TRUE, ...)

earea_(p, serie, name = NULL, stack = NULL, smooth = TRUE, ...)
}
\arguments{
\item{p}{an echart object.}

\item{serie}{value column name to plot.}

\item{name}{of serie.}

\item{stack}{name of the stack.}

\item{smooth}{whether to smooth line.}

\item{...}{any other argument to pass to the serie. i.e.: same parameters as \code{\link{eline}} or \code{\link{eline_}}}
}
\description{
Add area

Add area serie.
}
\examples{
mtcars \%>\%
  echart(mpg) \%>\%
  earea(qsec)

mtcars \%>\%
  echart(disp) \%>\%
  ebar(mpg) \%>\%
  earea(qsec, stack = "grp") \%>\%
  earea(wt, stack = "grp") \%>\%
  etooltip()

mtcars \%>\%
  echart_("mpg") \%>\%
  earea_("qsec")

mtcars \%>\%
  echart_("disp") \%>\%
  earea_("mpg", stack = "grp") \%>\%
  earea_("qsec", stack = "grp") \%>\%
  earea_("wt", stack = "grp") \%>\%
  etooltip() \%>\%
  elegend()

}

