% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{ebar_}
\alias{ebar_}
\title{Add bars}
\usage{
ebar_(p, serie, name = NULL, stack = NULL, clickable = TRUE,
  xAxisIndex = 0, yAxisIndex = 0, barGap = "100\%",
  barCategoryGap = "20\%", legendHoverLink = TRUE, z = 2, zlevel = 0,
  tooltip, ...)
}
\description{
Add bar serie.
}
\examples{
mtcars \%>\%
  echart_("mpg") \%>\%
  ebar_("qsec")

mtcars \%>\%
  echart_("disp") \%>\%
  ebar_("mpg", stack = "grp") \%>\% # stack
  ebar_("qsec", stack = "grp") \%>\% # stack
  ebar_("wt") \%>\% # not stacked
  etooltip() \%>\%
  elegend()

}

