% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarts.R
\name{echart}
\alias{echart}
\alias{echart_}
\title{Initiate an echart}
\usage{
echart(data, x, width = NULL, height = NULL, elementId = NULL)

echart_(data, x, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{data.frame containing data to plot.}

\item{x}{variable column.}

\item{width, height}{dimensions of chart.}

\item{elementId}{id of div containing chart.}
}
\description{
Initiate an echart graph.
}
