% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarts.R
\name{echarts-shiny}
\alias{echarts-shiny}
\alias{echartsOutput}
\alias{renderEcharts}
\alias{echartsProxy}
\title{Shiny bindings for echarts}
\usage{
echartsOutput(outputId, width = "100\%", height = "400px")

renderEcharts(expr, env = parent.frame(), quoted = FALSE)

echartsProxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a echarts}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{id}{target chart id.}

\item{session}{shiny session}
}
\description{
Output and render functions for using echarts within Shiny
applications and interactive Rmd documents.
}
