% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{echord}
\alias{echord}
\title{Add chord}
\usage{
echord(p, sort = "ascending", sortSub = "descending", ...)
}
\description{
Add chord
}
\examples{
set.seed(19880525)
matrix <- matrix(sample(0:1, 100, replace=TRUE, prob=c(0.9,0.6)), nc=10)

matrix \%>\%
  echart(LETTERS[1:10]) \%>\%
  echord()

}

