% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{echord_}
\alias{echord_}
\title{Add chord}
\usage{
echord_(p, name = NULL, sort = "none", sortSub = "none",
  clickable = TRUE, z = 2, zlevel = 0, symbol = NULL,
  symbolSize = NULL, clockWise = FALSE, minRadius = 10, maxRadius = 20,
  ribbonType = TRUE, showScale = FALSE, showScaleText = FALSE,
  padding = 2, ...)
}
\description{
Add chord
}
\examples{
set.seed(19880525)
matrix <- matrix(sample(0:1, 100, replace=TRUE, prob=c(0.9,0.6)), nc=10)

matrix \%>\%
  echart(LETTERS[1:10]) \%>\%
  echord()

matrix \%>\%
  echart(LETTERS[1:10]) \%>\%
  echord(ribbonType = FALSE)

}

