% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{echoropleth}
\alias{echoropleth}
\title{Add choropleth}
\usage{
echoropleth(p, serie, mapType = "world", dataRange, ...)
}
\description{
Add choropleth
}
\examples{
choropleth <- data.frame(countries = c("France", "Brazil", "China", "Russia", "Canada", "India"),
  values = runif(6, 10, 25))

choropleth \%>\%
  echart(countries) \%>\%
  echoropleth(values)

}

