% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{ecloud}
\alias{ecloud}
\title{add wordcloud}
\usage{
ecloud(p, freq, color, ...)
}
\description{
add wordcloud
}
\examples{
tf <- data.frame(terms = c("ECharts", "htmlwidgets", "rstats", "htmltools"),
  freq = c(20, 17, 15, 7), color = c("Red", "orange", "yellow", "grey"))

tf \%>\%
  echart(terms) \%>\%
  ecloud(freq, color)

}

