% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ecolorbar}
\alias{ecolorbar}
\title{Customise colorbar}
\usage{
ecolorbar(p, min = NULL, max = NULL, which = "previous", show = TRUE,
  color = NULL, zlevel = 4, z = 0, orient = "vertical", x = "left",
  y = "bottom", backgroundColor = "rgba(0,0,0,0)", borderColor = "#ccc",
  borderWidth = 0, padding = 5, itemGap = 10, itemWidth = 20,
  itemHeight = 14, precision = 0, splitNumber = 5, splitList = NULL,
  range = NULL, selectedMode = TRUE, calculable = FALSE,
  hoverLink = TRUE, realtime = FALSE, ...)
}
\arguments{
\item{p}{an echart object.}

\item{min, max}{minimum and maximum.}

\item{which}{series to serie is to be affected, takes the name of a serie, \code{previous} or \code{all}.}

\item{show}{whether to show the color bar.}

\item{color}{colors as list from high to low. i.e.: \code{list("red", "blue")}.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{orient}{orientation of bar, \code{vertical} or \code{horizontal}.}

\item{x}{x position; \code{left} or \code{right}.}

\item{y}{y posotion; \code{top} or \code{bottom}.}

\item{backgroundColor}{background color.}

\item{borderColor}{border color.}

\item{borderWidth}{width of border.}

\item{padding}{padding.}

\item{itemGap}{gap between items on bar.}

\item{itemWidth}{width of the bar.}

\item{itemHeight}{height of the bar.}

\item{precision}{decimal precision.}

\item{splitNumber}{number of segments.}

\item{splitList}{see \href{http://echarts.baidu.com/echarts2/doc/option-en.html#dataRange.splitList}{official docs} for details.}

\item{range}{used to set initial range i.e.: \code{list(start = 10, end = 50)}.}

\item{selectedMode}{selection mode.}

\item{calculable}{whether values are calculable.}

\item{hoverLink}{hoverlink with map.}

\item{realtime}{set to \code{TRUE} if using real time stream.}

\item{...}{any other argument to pass to color bar.}
}
\description{
Customise the colorbar of your chart.
}
\details{
\code{ecolorbar} refers to \href{http://echarts.baidu.com/echarts2/doc/option-en.html#dataRange.hoverLink}{datarange} in docs.
}
\examples{
df <- data.frame(x = 1:20,
                 y = runif(20, 5, 10),
                 size = runif(20, 5, 15))

df \%>\%
  echart(x) \%>\%
  escatter(y, size, symbolSize = 10, legendHoverLink = TRUE) \%>\%
  ecolorbar(color = list("red", "blue"), min = 5, max = 15, calculable = TRUE)

}
\seealso{
\href{http://echarts.baidu.com/echarts2/doc/option-en.html#dataRange.hoverLink}{official dataRange docs}
}
