% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{edata}
\alias{edata}
\alias{edata}
\alias{edata_}
\title{Add data}
\usage{
edata(p, data, x)

edata_(p, data, x)
}
\arguments{
\item{p}{an echart object.}

\item{data}{data.frame.}

\item{x}{x variable.}
}
\description{
Add a dataset.
}
\seealso{
\code{emap}
}
