% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{edatarange}
\alias{edatarange}
\title{Customise data range}
\usage{
edatarange(p, min = NULL, max = NULL, which = "previous", show = TRUE,
  color = list("#1e90ff", "#f0ffff"), zlevel = 4, z = 0,
  orient = "vertical", x = "left", y = "bottom",
  backgroundColor = "rgba(0,0,0,0)", borderColor = "#ccc",
  borderWidth = 0, padding = 5, itemGap = 10, itemWidth = 20,
  itemHeight = 14, precision = 0, splitNumber = 5, splitList = NULL,
  range = NULL, selectedMode = TRUE, calculable = FALSE,
  hoverLink = TRUE, realtime = FALSE, text = list("High", "Low"), ...)
}
\description{
Customise data range
}

