% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{efunnel}
\alias{efunnel}
\title{Add funnel}
\usage{
efunnel(p, serie, ...)
}
\description{
Add funnel
}
\examples{
funnel <- data.frame(stage = c("View", "Click", "Purchase"), value = c(80, 30, 20))

funnel \%>\%
  echart(stage) \%>\%
  efunnel(value)

}

