% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{efunnel_}
\alias{efunnel_}
\title{Add funnel}
\usage{
efunnel_(p, serie, name = NULL, clickable = TRUE, legendHoverLink = TRUE,
  sort = "descending", min = 0, max = 100, x = 80, y = 60, x2 = 80,
  y2 = 60, width = NULL, height = NULL, funnelAlign = "center",
  minSize = "0\%", maxSize = "100\%", gap = 0, tooltip, ...)
}
\description{
Add funnel
}
\examples{
funnel <- data.frame(stage = c("View", "Click", "Purchase"), value = c(80, 30, 20))

funnel \%>\%
  echart_("stage") \%>\%
  efunnel_("value")

}

