% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{egauge}
\alias{egauge}
\alias{egauge_}
\title{Add gauge}
\usage{
egauge(p, value, indicator = "", name = NULL, clickable = FALSE,
  legendHoverLink = TRUE, center = list("50\%", "50\%"),
  radius = list("0\%", "75\%"), startAngle = 225, endAngle = -45,
  min = 0, max = 100, splitNumber = 10, z = 2, zlevel = 0, tooltip,
  ...)

egauge_(p, value, indicator = "", name = NULL, clickable = FALSE,
  legendHoverLink = TRUE, center = list("50\%", "50\%"),
  radius = list("0\%", "75\%"), startAngle = 225, endAngle = -45,
  min = 0, max = 100, splitNumber = 10, z = 2, zlevel = 0, tooltip,
  ...)
}
\arguments{
\item{p}{an echart object.}

\item{value}{value to plot.}

\item{indicator}{indicator appearing in center of gauge.}

\item{name}{name of serie.}

\item{clickable}{whether the item is clickable.}

\item{legendHoverLink}{enables legend hover link.}

\item{center}{center of gauge in pixels of percent.}

\item{radius}{radius of gauge in pixels of percent.}

\item{startAngle, endAngle}{start and end angles of gauge.}

\item{min, max}{minimum and maximum of gauge.}

\item{splitNumber}{number of segments.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{tooltip}{customise tooltip.}

\item{...}{any other arguments.}
}
\description{
Add gauge.
}
\examples{
echart() \%>\%
  egauge(85, "SPEED")

echart() \%>\%
  egauge(25, "SPEED") \%>\%
  etheme("helianthus")

echart() \%>\%
  egauge(63, "PERCENT") \%>\%
  etheme("dark")

}
\seealso{
\href{http://echarts.baidu.com/echarts2/doc/option-en.html#series-i(gauge)}{official gauge docs}
}

