% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{egauge_}
\alias{egauge_}
\title{Add gauge}
\usage{
egauge_(p, value, indicator = "", name = NULL, clickable = TRUE,
  legendHoverLink = TRUE, center = list("50\%", "50\%"),
  radius = list("0\%", "75\%"), startAngle = 225, endAngle = -45,
  min = 0, max = 100, splitNumber = 10, z = 2, zlevel = 0, tooltip,
  ...)
}
\description{
Add gauge
}
\examples{
echart() \%>\%
  egauge(85, "SPEED")

echart() \%>\%
  egauge(85, "SPEED") \%>\%
  etheme("helianthus")

}

