% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{egrid}
\alias{egrid}
\title{Customise grid}
\usage{
egrid(p, backgroundColor = NULL, borderWidth = 1, borderColor = NULL,
  width = NULL, height = NULL, z = 0, zlevel = 0, x = 80, y = 60,
  x2 = 80, y2 = 80)
}
\arguments{
\item{p}{an echart object.}

\item{backgroundColor}{background color.}

\item{borderWidth}{border width.}

\item{borderColor}{border color.}

\item{width, height}{dimensions of grid.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{x, y}{ordinate on upper left corner.}

\item{x2, y2}{ordinate on upper right corner.}
}
\description{
Customise grid
}
\examples{
df <- data.frame(x = 1:20, y = runif(20, 5, 20))

df \%>\%
  echart(x) \%>\%
  eline(x) \%>\%
  egrid(borderWidth = 5, borderColor = "red", backgroundColor = "yellow")

}

