% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{eheatmap}
\alias{eheatmap}
\alias{eheatmap}
\alias{eheatmap_}
\title{Add heatmap}
\usage{
eheatmap(p, y, values, name = NULL, clickable = TRUE, blurSize = 30,
  minAlpha = 0.5, valueScale = 1, opacity = 1, z = 2, zlevel = 0,
  gradientColors, tooltip, ...)

eheatmap_(p, y, values, name = NULL, clickable = TRUE, blurSize = 30,
  minAlpha = 0.5, valueScale = 1, opacity = 1, z = 2, zlevel = 0,
  gradientColors, tooltip, ...)
}
\arguments{
\item{p}{an echart object.}

\item{y}{yaxis values.}

\item{values}{heat.}

\item{name}{name of serie.}

\item{clickable}{whether chart is clickable.}

\item{blurSize}{size of points blur.}

\item{minAlpha}{minimum transparency.}

\item{valueScale}{\code{values} multiplier.}

\item{opacity}{opacity of heatmap.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{gradientColors}{colors used for gradient as a \code{list} i.e.:\code{list("red", "blue")}}

\item{tooltip}{cutomise tooltip.}

\item{...}{any other options to pass to heatmap.}
}
\description{
Add heatmap.
}
\examples{
set.seed(19880525)
matrix <- data.frame(x = runif(150, 10, 500), y = runif(150, 10, 500), z = runif(150, 10 , 200))

matrix \%>\%
  echart_("x") \%>\%
  eheatmap_("y", "z")

}
\seealso{
\href{http://echarts.baidu.com/echarts2/doc/option-en.html#series-i(heatmap)}{official heatmap docs}
}
