% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{eheatmap_}
\alias{eheatmap_}
\title{add heatmap}
\usage{
eheatmap_(p, y, values, name = NULL, clickable = TRUE, blurSize = 30,
  minAlpha = 0.5, valueScale = 1, opacity = 1, z = 2, zlevel = 0,
  gradientColors, tooltip, ...)
}
\description{
add heatmap
}
\examples{
set.seed(19880525)
matrix <- data.frame(x = runif(150, 10, 500), y = runif(150, 10, 500), z = runif(150, 10 , 200))

matrix \%>\%
  echart_("x") \%>\%
  eheatmap_("y", "z")

}

