% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{elegend}
\alias{elegend}
\title{Add legend}
\usage{
elegend(p, legend, show = TRUE, zlevel = 0, z = 4,
  orient = "horizontal", x = "center", y = "top",
  backgroundColor = "rgba(0,0,0,0)", borderColor = "#ccc",
  borderWidth = 0, padding = 5, itemGap = 10, itemWidth = 20,
  itemHeight = 14, formatter = NULL, selectedMode = TRUE,
  selected = NULL, textStyle, ...)
}
\description{
Add legend
}

