% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{eline}
\alias{eline}
\title{Add lines}
\usage{
eline(p, serie, fill = FALSE, stack = NULL, ...)
}
\description{
Add line serie.
}
\examples{
mtcars \%>\%
  echart(mpg) \%>\%
  eline(qsec)

mtcars \%>\%
  echart(disp) \%>\%
  eline(mpg, stack = "grp") \%>\%
  eline(qsec, stack = "grp") \%>\%
  eline(wt, fill = TRUE) \%>\%
  etooltip()

}

