% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{eline_}
\alias{eline_}
\title{Add lines}
\usage{
eline_(p, serie, name = NULL, stack = NULL, clickable = TRUE,
  xAxisIndex = 0, yAxisIndex = 0, symbol = NULL, symbolSize = "2 | 4",
  symbolRate = NULL, showAllSymbol = FALSE, smooth = TRUE,
  legendHoverLink = TRUE, dataFilter = "nearest", z = 2, zlevel = 0,
  tooltip, ...)
}
\description{
Add line serie.
}
\examples{
mtcars \%>\%
  echart_("mpg") \%>\%
  eline_("qsec")

mtcars \%>\%
  echart_("disp") \%>\%
  eline_("mpg", stack = "grp") \%>\%
  eline_("qsec", stack = "grp") \%>\%
  eline_("wt", fill = TRUE) \%>\%
  etooltip() \%>\%
  elegend() \%>\%
  etoolbox_magic(type = list("line", "bar"))

}

