% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{eload}
\alias{eload}
\title{Data loading options}
\usage{
eload(p, text = "loading", effect = "bubble", x = "center",
  y = "center", progress = NULL, ...)
}
\arguments{
\item{p}{an echart object.}

\item{text}{the loading text that appears when the chart is set into the loading state, use \code{\\n} for new line.}

\item{effect}{loading effect, see details for valid values.}

\item{x, y}{position, takes \code{left}, \code{right}, \code{center}.}

\item{progress}{specifies the current progress \code{[0~1]}. Applicable to specific effects.}

\item{...}{any other parameters, i.e.:\code{textStyle}.}
}
\description{
Customise the data loading options.
}
\details{
Valid \code{effect} values:
\itemize{
  \item{spin}
  \item{bar}
  \item{ring}
  \item{whirling}
  \item{dynamicLine}
  \item{bubble}
}
}
\examples{
# no data
df <- data.frame(x = NA, y = NA) \%>\%
  dplyr::filter(!is.na(x))

df \%>\%
  echart(x) \%>\%
  eline(y)

df \%>\%
  echart(x) \%>\%
  eline(y) \%>\%
  eload(effect = "ring", text = "loading data...")

df \%>\%
  echart(x) \%>\%
  eline(y) \%>\%
  eload(effect = "dynamicLine")

}
