% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{emap_choropleth}
\alias{emap_choropleth}
\title{Add choropleth}
\usage{
emap_choropleth(p, serie, dataRange = NULL)
}
\description{
Add choropleth
}
\examples{
choropleth <- data.frame(countries = c("France", "Brazil", "China", "Russia", "Canada", "India"),
  values = round(runif(6, 10, 25)))

choropleth \%>\%
  echart(countries) \%>\%
  emap() \%>\%
  emap_choropleth(values)

}

