% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{emap_choropleth_}
\alias{emap_choropleth_}
\title{Add choropleth}
\usage{
emap_choropleth_(p, serie, dataRange = NULL)
}
\description{
Add choropleth
}
\examples{
choropleth <- data.frame(countries = c("France", "Brazil", "China", "Russia", "Canada", "India"),
  values = round(runif(6, 10, 25)))

choropleth \%>\%
  echart_("countries") \%>\%
  emap() \%>\%
  emap_choropleth_("values")

}

