% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{emap_heat}
\alias{emap_heat}
\alias{emap_heat_}
\title{Add heat on map}
\usage{
emap_heat(p, lon, lat, z, blurSize = 30, minAlpha = 0.05, valueScale = 1,
  opacity = 1, gradientColors = NULL, ...)

emap_heat_(p, lon, lat, z, blurSize = 30, minAlpha = 0.05, valueScale = 1,
  opacity = 1, gradientColors = NULL, ...)
}
\arguments{
\item{p}{an echart object.}

\item{lon, lat}{coordinates.}

\item{z}{values, heat.}

\item{blurSize}{blur of points.}

\item{minAlpha}{minimum transparency.}

\item{valueScale}{\code{z} multiplier.}

\item{opacity}{opacity of heatmap.}

\item{gradientColors}{colors.}

\item{...}{any other parameter to pass to heatmap.}
}
\description{
Add heat on map
}
\examples{
data <- data.frame(lon = runif(300, 90, 120),
  lat = runif(300, 30, 39),
  z = runif(300, 75, 100))

data \%>\%
  echart_() \%>\%
  emap(mapType = "china") \%>\%
  emap_heat_("lon", "lat", "z")

data \%>\%
  echart() \%>\%
  emap(mapType = "china") \%>\%
  emap_heat_("lon", "lat", "z", blurSize = 50, minAlpha = 0.3, opacity = 0.8)

}

