% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{emap_line_effect}
\alias{emap_line_effect}
\title{emap line effect}
\usage{
emap_line_effect(show = TRUE, loop = TRUE, period = 30, scaleSize = 1,
  color = "#fff", shadowBlur = 10, shadowColor = NULL, ...)
}
\arguments{
\item{show}{set to \code{TRUE} to show effect.}

\item{loop}{set to \code{TRUE} to loop animation.}

\item{period}{period loop.}

\item{scaleSize}{scale.}

\item{color}{color.}

\item{shadowBlur}{blur.}

\item{shadowColor}{color of shadow.}

\item{...}{any other option to pass to effect.}
}
\description{
Effect for emap lines
}
\examples{
coords <- data.frame(city = c("London", "New York", "Beijing", "Sydney"),
  lon = c(-0.1167218, -73.98002, 116.3883, 151.18518),
  lat = c(51.49999, 40.74998, 39.92889, -33.92001))

edges <- data.frame(source = c("Beijing", "Beijing", "New York"),
  target = c("Sydney", "London", "London"))

coords \%>\%
  echart(city) \%>\%
  emap() \%>\%
  emap_coords(lon, lat) \%>\%
  emap_lines(edges, source, target, effect = emap_line_effect())

}
