% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{emap_lines_}
\alias{emap_lines_}
\title{Add map lines}
\usage{
emap_lines_(p, edges, source, target, name = NULL, clickable = TRUE,
  symbol = list("circle", "arrow"), symbolSize = list(2, 4),
  symbolRate = NULL, large = FALSE, smooth = TRUE, z = 2, zlevel = 0,
  smoothness = 0.2, precision = 2, bundling = list(enable = FALSE,
  maxTurningAngle = 45), ...)
}
\description{
Add lines on map
}
\examples{
coords <- data.frame(city = c("London", "New York", "Beijing", "Sydney"),
  lon = c(-0.1167218, -73.98002, 116.3883, 151.18518),
  lat = c(51.49999, 40.74998, 39.92889, -33.92001))

edges <- data.frame(source = c("Beijing", "Beijing", "New York"),
  target = c("Sydney", "London", "London"))

coords \%>\%
  echart_("city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_lines_(edges, "source", "target")

edges2 <- data.frame(source = "London", target = "Sydney")

coords \%>\%
  echart_("city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_lines_(edges, "source", "target") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_lines_(edges2, "source", "target", effect = emap_line_effect()) \%>\%
  etheme("helianthus")

coords2 <- data.frame(city = "Sydney", lon = 151.18518, lat = -33.92001, value = 20)

coords \%>\%
  echart_("city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_lines_(edges, "source", "target") \%>\%
  edata_(coords2, "city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_lines_(edges2, "source", "target", effect = emap_line_effect(scaleSize = 2)) \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_points_("value", symbol = "emptyCircle", effect = list(show = TRUE, shadowBlur = 10)) \%>\%
  etheme("dark")

}
\seealso{
\code{\link{emap_coords}}
}

