% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{emap_points}
\alias{emap_points}
\title{Add map points}
\usage{
emap_points(p, serie)
}
\description{
Add map points
}
\examples{
coords <- data.frame(city = c("London", "New York", "Beijing", "Sydney"),
  lon = c(-0.1167218, -73.98002, 116.3883, 151.18518),
  lat = c(51.49999, 40.74998, 39.92889, -33.92001),
  values = runif(4, 10, 20))

coords \%>\%
  echart(city) \%>\%
  emap_coords(lon, lat) \%>\%
  emap_points(values)

}

