% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{emap_points_}
\alias{emap_points_}
\title{Add map points}
\usage{
emap_points_(p, serie, clickable = TRUE, symbol = "pin",
  symbolSize = htmlwidgets::JS(" function (v){ return 10 + v/10 }"),
  symbolRotate = NULL, large = FALSE, itemStyle, effect, ...)
}
\description{
Add map points
}
\examples{
coords <- data.frame(city = c("London", "New York", "Beijing", "Sydney"),
  lon = c(-0.1167218, -73.98002, 116.3883, 151.18518),
  lat = c(51.49999, 40.74998, 39.92889, -33.92001),
  values = runif(4, 10, 20))

coords \%>\%
  echart(city) \%>\%
  emap() \%>\%
  emap_coords(lon, lat) \%>\%
  emap_points(values)

coords2 <- data.frame(city = "Rio", lon = -43.172896, lat = -22.906847, value = 15)

coords \%>\%
  echart(city) \%>\%
  emap() \%>\%
  emap_coords(lon, lat) \%>\%
  emap_points(values, symbolSize = 5) \%>\%
  edata(coords2, city) \%>\%
  emap() \%>\%
  emap_coords(lon, lat) \%>\%
  emap_points(value, symbol = "emptyCircle", effect = list(show = TRUE, shadowBlur = 10)) \%>\%
  etheme("infographic")

}

