% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{emap_roam}
\alias{emap_roam}
\title{Add Zoom and roam controller}
\usage{
emap_roam(p, show = TRUE, zlevel = 0, z = 4, x = "left", y = "top",
  width = 80, height = 120, backgroundColor = "rgba(0,0,0,0)",
  borderColor = "#ccc", borderWidth = 0, padding = 5,
  fillerColor = "#fff", handleColor = "#6495ed", step = 15,
  mapTypeControl = NULL, ...)
}
\arguments{
\item{p}{an echart object.}

\item{show}{set to \code{TRUE} to show the controller}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{x}{x position; \code{left} or \code{right}.}

\item{y}{y posotion; \code{top} or \code{bottom}.}

\item{width, height}{dimensions of controller.}

\item{backgroundColor}{background color.}

\item{borderColor}{border color.}

\item{borderWidth}{width of border.}

\item{padding}{padding.}

\item{fillerColor}{filler color.}

\item{handleColor}{color of handle.}

\item{step}{moving step of the 4 direction roam in px.}

\item{mapTypeControl}{ou can specify every single mapType when multiple map in a chart at the same time, such as: \code{list({ china = FALSE, world = TRUE})}.}

\item{...}{any other option to pass to controller.}
}
\description{
Add zoom and roam controller to map.
}
\examples{
coords <- data.frame(city = c("London", "New York", "Beijing", "Sydney"),
  lon = c(-0.1167218, -73.98002, 116.3883, 151.18518),
  lat = c(51.49999, 40.74998, 39.92889, -33.92001),
  values = runif(4, 10, 20))

coords \%>\%
  echart_("city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_points_("values") \%>\%
  emap_roam(mapTypeControl = list(world = TRUE))

}

