% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{emark_line}
\alias{emark_line}
\title{mark line}
\usage{
emark_line(p, which = "previous", data = list(), clickable = TRUE,
  symbol = list("circle", "arrow"), symbolSize = list(2, 4),
  symbolRotate = NULL, large = FALSE, smooth = FALSE, precision = 2,
  bundling = list(enable = FALSE, maxTurningAngle = 45), z = 2,
  zlevel = 0, effect, itemStyle, ...)
}
\description{
mark line
}

