% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{emark_point}
\alias{emark_point}
\title{mark points}
\usage{
emark_point(p, which = "previous", data = list(), clickable = TRUE,
  symbol = "pin", symbolSize = 10, symbolRotate = NULL, large = FALSE,
  effect, itemStyle, ...)
}
\description{
mark points
}

