% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_.R
\name{enodes_}
\alias{enodes_}
\title{Add nodes}
\usage{
enodes_(p, nodes, name, label = NULL, value = NULL, category = NULL,
  symbolSize = NULL, depth = NULL, ignore = FALSE, symbol = "circle",
  fixX = FALSE, fixY = FALSE)
}
\description{
Add nodes for \code{\link{eforce}}.
}
\examples{
let <- LETTERS[1:20]

edges <- data.frame(source = sample(let, 20), target = sample(let, 20),
  weight = runif(20, 5, 20))

nodes <- data.frame(name = let, value = runif(20, 5, 25), group = rep(LETTERS[1:4], 5))

echart() \%>\%
  eforce_(itemStyle = list(normal = list(label = list(show = TRUE)))) \%>\% # show labels
  enodes_(nodes, "name", value = "value", category = "group") \%>\%
  elinks_(edges, "source", "target")

}

