% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{epie}
\alias{epie}
\title{Add pie}
\usage{
epie(p, serie, name = NULL, ...)
}
\description{
Add pie chart
}
\examples{
pie <- data.frame(name = c("banana", "apple", "pineapple", "onion"),
  value = c(40, 15, 12, 9))

pie \%>\%
  echart(name) \%>\%
  epie(value)

pie \%>\%
  echart(name) \%>\%
  epie(value, roseType = "area") \%>\%
  etheme("helianthus")

pie \%>\%
  echart(name) \%>\%
  epie(value, roseType = "radius") \%>\%
  etheme("mint")

}

