% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{epie_}
\alias{epie_}
\title{Add pie}
\usage{
epie_(p, serie, name = NULL, ...)
}
\description{
Add pie chart
}
\examples{
pie <- data.frame(name = c("banana", "apple", "pineapple", "onion"),
  value = c(26, 15, 12, 9))

pie \%>\%
  echart_("name") \%>\%
  epie_("value")

pie \%>\%
  echart_("name") \%>\%
  epie_("value", roseType = "area") \%>\%
  etheme("helianthus")

pie \%>\%
  echart_("name") \%>\%
  epie_("value", roseType = "radius") \%>\%
  etheme("mint")

}

