% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{eradar}
\alias{eradar}
\title{add radar}
\usage{
eradar(p, serie, name = NULL, ...)
}
\description{
add radar
}
\examples{
radar <- data.frame(axis = LETTERS[1:6], value = runif(6, 2, 10))

radar \%>\%
  echart(axis) \%>\%
  eradar(value)

}

