% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{eradar_}
\alias{eradar_}
\title{add radar}
\usage{
eradar_(p, serie, name = NULL, ...)
}
\description{
add radar
}
\examples{
radar <- data.frame(axis = rep(LETTERS[1:6], 4), grp = rep(LETTERS[4:9], 4),
  value = runif(24, 2, 10))

radar \%>\%
  group_by_("grp") \%>\%
  echart_("axis") \%>\%
  eradar_("value") \%>\%
  elegend()

}

