% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{escatter}
\alias{escatter}
\title{Add scatter}
\usage{
escatter(p, serie, size = NULL, name = NULL, clickable = TRUE,
  symbol = NULL, symbolSize = 4, symbolRotate = NULL, large = FALSE,
  largeThreshold = 2000, legendHoverLink = TRUE, z = 2, zlevel = 0, ...)
}
\description{
Add scatter serie.
}
\examples{
mtcars \%>\%
  echart(disp) \%>\%
  escatter(mpg)

mtcars \%>\%
  echart(disp) \%>\%
  escatter(mpg, qsec)

}

