% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{etheme}
\alias{etheme}
\title{Add theme}
\usage{
etheme(p, theme = "default")
}
\arguments{
\item{p}{an echart object.}

\item{theme, }{see details for valid values.}
}
\description{
Add a theme.
}
\details{
valid themes:
\itemize{
  \item{\code{default}}
  \item{\code{mint}}
  \item{\code{macarons}}
  \item{\code{macarons2}}
  \item{\code{green}}
  \item{\code{blue}}
  \item{\code{dark}}
  \item{\code{gray}}
  \item{\code{helianthus}}
  \item{\code{red}}
  \item{\code{roma}}
  \item{\code{sakura}}
  \item{\code{shine}}
  \item{\code{infographic}}
}
}
\examples{
mtcars \%>\%
  echart(disp) \%>\%
  ebar(qsec) \%>\%
  ebar(mpg) \%>\%
  etheme("roma")

}

