% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{etitle}
\alias{etitle}
\title{Add title}
\usage{
etitle(p, text, subtext, link, sublink, target = "blank",
  subtarget = "blank", x = "left", y = "top",
  backgroundColor = "rgba(0,0,0,0)", borderColor = "#ccc",
  borderWidth = 0, padding = 5, itemGap = 5, zlevel = 0, z = 6,
  show = TRUE, ...)
}
\description{
Add title
}

