% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{etoolbox_feature}
\alias{etoolbox_feature}
\title{Add toolbox feature}
\usage{
etoolbox_feature(p, mark, dataZoom, dataView, magicType, restore, saveAsImage)
}
\arguments{
\item{p}{an echart object.}

\item{mark}{markLine icons see \code{\link{etoolbox_mark}}.}

\item{dataZoom}{dataZoom icons \code{\link{etoolbox_zoom}}.}

\item{dataView}{dataView icons \code{\link{etoolbox_view}}.}

\item{magicType}{magicType icons \code{\link{etoolbox_magic}}.}

\item{restore}{restore icon \code{\link{etoolbox_restore}}.}

\item{saveAsImage}{saveAsImage icon \code{\link{etoolbox_save}}.}
}
\description{
Add toolbox feature.
}
\examples{
mtcars \%>\%
  echart(qsec) \%>\%
  ebar(mpg) \%>\%
  etoolbox() \%>\%
  etoolbox_magic(type = list("line", "bar")) \%>\%
  etoolbox_feature(restore = list(show = TRUE))

}

