% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{etoolbox_full}
\alias{etoolbox_full}
\title{Add all elements of toolbox}
\usage{
etoolbox_full(p, ...)
}
\arguments{
\item{p}{an echart object.}

\item{...}{any other option to pass to \code{\link{etoolbox}}.}
}
\description{
Adds toolbok mark, restor, save, and view.
}
\details{
Adds mark, restore, save, view and zoom buttons
}
