% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{etoolbox_view}
\alias{etoolbox_view}
\title{Add toolbox view}
\usage{
etoolbox_view(p, show = TRUE, title = "View", readOnly = FALSE,
  lang = list("Data View", "close", "refresh"), ...)
}
\description{
Add toolbox view
}

