% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{etooltip}
\alias{etooltip}
\title{Add tooltip}
\usage{
etooltip(p, show = TRUE, trigger = "axis", zlevel = 1, z = 8,
  showContent = TRUE, position = NULL, formatter = NULL,
  islandFormatter = "{a} < br/>{b} : {c}", showDelay = 20,
  hideDelay = 100, transitionDuration = 4, enterable = FALSE,
  backgroundColor = "rgba(0,0,0,0.7)", borderColor = "#333",
  borderRadius = 4, borderWidth = 0, padding = 5, axisPointer, textStyle,
  ...)
}
\description{
Add tooltip
}

