% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{etreemap}
\alias{etreemap}
\alias{etreemap}
\alias{etreemap_}
\title{Add Treemap}
\usage{
etreemap(p, serie, name = NULL, itemStyle = NULL, clickable = FALSE,
  center = list("50\%", "50\%"), size = list("80\%", "80\%"), z = 2,
  zlevel = 0, ...)

etreemap_(p, serie, name = NULL, itemStyle = NULL, clickable = FALSE,
  center = list("50\%", "50\%"), size = list("80\%", "80\%"), z = 2,
  zlevel = 0, ...)
}
\arguments{
\item{p}{an echart object.}

\item{serie}{values to plot.}

\item{name}{name of serie.}

\item{itemStyle}{style of rectangles.}

\item{clickable}{whether rectangles are clickable.}

\item{center}{center of map.}

\item{size}{size of chart.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{...}{any other option to pass to treemap.}
}
\description{
Add Treemap
}
\examples{
df <- data.frame(name = LETTERS[1:10], values = round(runif(10, 1, 10)))

df \%>\%
  echart(name) \%>\%
  etreemap(values) \%>\%
  etooltip(trigger = "item") \%>\%
  etheme("macarons")

}
