% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{evenn}
\alias{evenn}
\title{Add venn}
\usage{
evenn(p, serie, overlap, ...)
}
\arguments{
\item{serie}{a named vector, see details.}
}
\description{
Add venn diagram
}
\examples{
venn <- data.frame(name = c("banana", "pineapple", "overlap"),
  values = c(20, 50, 10))

venn \%>\%
  echart(name) \%>\%
  evenn(values)

}

