% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{evenn_}
\alias{evenn_}
\title{Add venn}
\usage{
evenn_(p, serie, overlap, name = NULL, clickable = TRUE, z = 2,
  zlevel = 0, tooltip, itemStyle, markPoint, markLine, ...)
}
\arguments{
\item{serie}{a named vector, see details.}
}
\description{
Add venn diagram
}
\examples{
venn <- data.frame(name = c("banana", "pineapple", "overlap"),
  values = c(20, 50, 10))

venn \%>\%
  echart(name) \%>\%
  evenn(values)

}

