% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_.R
\name{igraph2eforce}
\alias{igraph2eforce}
\alias{igraph_nodes}
\alias{igraph_links}
\title{Plot igraph}
\usage{
igraph_nodes(g)

igraph_links(g)
}
\arguments{
\item{g}{igraph object}
}
\description{
Plot a network from an igraph object.
}
\examples{
g <- igraph::make_directed_graph(LETTERS)

nodes <- igraph_nodes(g)
links <- igraph_links(g)

echart() \%>\%
  eforce_() \%>\%
  enodes_(nodes, "name", ) \%>\%
  elinks_(links, "source", "target")

}
