% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{\%>\%}
\alias{group_by}
\alias{group_by_}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_by_}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}
\keyword{internal}

