% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{xAxis}
\alias{exAxis}
\alias{exAxis_category}
\alias{exAxis_log}
\alias{exAxis_time}
\alias{exAxis_value}
\alias{xAxis}
\title{Customize X axis}
\usage{
exAxis(p, show = TRUE, type = "value", ...)

exAxis_category(p, show = TRUE, zlevel = 0, z = 0, boundaryGap = FALSE,
  append = FALSE, ...)

exAxis_value(p, show = TRUE, zlevel = 0, z = 0, position = "bottom",
  name = NULL, nameLocation = "end", nameTextStyle = list(),
  boundaryGap = list(0, 0), min = NULL, max = NULL, scale = FALSE,
  splitNumber = NULL, append = FALSE, ...)

exAxis_time(p, show = TRUE, zlevel = 0, z = 0, position = "bottom",
  name = NULL, nameLocation = "end", nameTextStyle = list(),
  boundaryGap = list(0, 0), min = NULL, max = NULL, scale = FALSE,
  splitNumber = NULL, append = FALSE, ...)

exAxis_log(p, show = TRUE, zlevel = 0, z = 0, position = "bottom",
  logLabelBase = NULL, logPositive = NULL, append = FALSE, ...)
}
\description{
Customize X axis
}

