% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{yAxis}
\alias{eyAxis}
\alias{eyAxis_category}
\alias{eyAxis_log}
\alias{eyAxis_time}
\alias{eyAxis_value}
\alias{yAxis}
\title{Customise axis}
\usage{
eyAxis(p, show = TRUE, type = "value", append = FALSE, ...)

eyAxis_category(p, show = TRUE, zlevel = 0, z = 0, boundaryGap = FALSE,
  append = FALSE, ...)

eyAxis_value(p, show = TRUE, zlevel = 0, z = 0, position = "left",
  name = NULL, nameLocation = "end", nameTextStyle = list(),
  boundaryGap = list(0, 0), min = NULL, max = NULL, scale = FALSE,
  splitNumber = NULL, append = FALSE, ...)

eyAxis_time(p, show = TRUE, zlevel = 0, z = 0, position = "left",
  name = NULL, nameLocation = "end", nameTextStyle = list(),
  boundaryGap = list(0, 0), min = NULL, max = NULL, scale = FALSE,
  splitNumber = NULL, append = FALSE, ...)

eyAxis_log(p, show = TRUE, zlevel = 0, z = 0, position = "left",
  logLabelBase = NULL, logPositive = NULL, append = FALSE, ...)
}
\description{
Customise axis
}
\examples{
df <- data.frame(x = c("Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"),
  y = runif(7, 1, 5))

df \%>\%
  echart(x) \%>\%
  eline(y) \%>\%
  exAxis_category(boundaryGap = FALSE)

df \%>\%
  echart(x) \%>\%
  ebar(y) \%>\%
  eyAxis_log()

}

