% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{earea}
\alias{earea}
\title{Add area}
\usage{
earea(p, serie, name = NULL, clickable = TRUE, legendHoverLink = TRUE,
  center = list("50\%", "50\%"), radius = list(0, "75\%"),
  startAngle = 90, minAngle = 0, clockWise = TRUE, roseType = NULL,
  selectedOffset = 10, selectedMode = FALSE, z = 2, zlevel = 0, ...)
}
\description{
Add area serie
}
\examples{
mtcars \%>\%
  echart(mpg) \%>\%
  earea(qsec)

mtcars \%>\%
  echart(disp) \%>\%
  ebar(mpg) \%>\%
  earea(qsec, stack = "grp") \%>\%
  earea(wt, stack = "grp") \%>\%
  etooltip()

}

