% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{earea}
\alias{earea}
\title{Add area}
\usage{
earea(p, serie, name = NULL, stack = NULL, ...)
}
\description{
Add area serie
}
\examples{
mtcars \%>\%
  echart(mpg) \%>\%
  earea(qsec)

mtcars \%>\%
  echart(disp) \%>\%
  ebar(mpg) \%>\%
  earea(qsec, stack = "grp") \%>\%
  earea(wt, stack = "grp") \%>\%
  etooltip()

}

