% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{earea}
\alias{earea}
\alias{earea}
\alias{earea_}
\title{Add area}
\usage{
earea(p, serie, name = NULL, stack = NULL, smooth = TRUE, ...)

earea_(p, serie, name = NULL, stack = NULL, smooth = TRUE, ...)
}
\arguments{
\item{p}{an echart object.}

\item{serie}{value column name to plot.}

\item{name}{of serie.}

\item{stack}{name of the stack.}

\item{smooth}{whether to smooth line.}

\item{...}{any other argument to pass to the serie. i.e.: same parameters as \code{\link{eline}} or \code{\link{eline_}}}
}
\description{
Add area serie.
}
\examples{
\dontrun{
df <- data.frame(x = LETTERS[1:10], y = runif(10, 30, 70), z = runif(10, 10, 50))

df \%>\%
  echart_("x") \%>\%
  earea_("y", smooth = FALSE, symbol = "emptyRectangle", symbolSize = 5)

df \%>\%
  echart(x) \%>\%
  earea(y, stack = "grp") \%>\%
  earea(z, stack = "grp") \%>\%
  etheme("roma")

df <- data.frame(x = 1:10, y = runif(10, 30, 70), z = runif(10, 10, 50))

df \%>\%
  echart(x) \%>\%
  earea(z, stack = "grp") \%>\%
  earea(y)
}

}
