% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{earea_}
\alias{earea_}
\title{Add area}
\usage{
earea_(p, serie, name = NULL, stack = NULL, ...)
}
\description{
Add area serie
}
\examples{
mtcars \%>\%
  echart_("mpg") \%>\%
  earea_("qsec")

mtcars \%>\%
  echart_("disp") \%>\%
  earea_("mpg", stack = "grp") \%>\%
  earea_("qsec", stack = "grp") \%>\%
  earea_("wt", stack = "grp") \%>\%
  etooltip() \%>\%
  elegend()

}

