% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{ebar}
\alias{ebar}
\alias{ebar}
\alias{ebar_}
\title{Add bars}
\usage{
ebar(p, serie, name = NULL, stack = NULL, clickable = TRUE,
  xAxisIndex = 0, yAxisIndex = 0, barGap = "100\%",
  barCategoryGap = "20\%", legendHoverLink = TRUE, z = 2, zlevel = 0,
  tooltip, ...)

ebar_(p, serie, name = NULL, stack = NULL, clickable = TRUE,
  xAxisIndex = 0, yAxisIndex = 0, barGap = "30\%",
  barCategoryGap = "20\%", legendHoverLink = TRUE, z = 2, zlevel = 0,
  tooltip, ...)
}
\arguments{
\item{p}{an echart object.}

\item{serie}{value column name to plot.}

\item{name}{of serie.}

\item{stack}{name of the stack.}

\item{clickable}{whether plot is clickable.}

\item{xAxisIndex, yAxisIndex}{axis indexes.}

\item{barGap, barCategoryGap}{distance between each bar.}

\item{legendHoverLink}{enables legend hover links.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{tooltip}{style of tooltip.}

\item{...}{any other argument to pass to the serie.}
}
\description{
Add bar serie.
}
\examples{
mtcars \%>\%
  echart_("mpg") \%>\%
  ebar_("qsec")

mtcars \%>\%
  echart_("disp") \%>\%
  ebar_("mpg", stack = "grp") \%>\% # stack
  ebar_("qsec", stack = "grp") \%>\% # stack
  ebar_("wt", stack = "grp2") \%>\% # not stacked
  etooltip(trigger = "item") \%>\%
  elegend() \%>\%
  etoolbox_magic(type = list("stack", "tiled")) \%>\%
  etoolbox_restore()

df <- data.frame(x = LETTERS[1:4], y = runif(4, 0, 20), z = runif(4, 10, 15), w = runif(4, 15, 30))

df \%>\%
  echart(x) \%>\%
  ebar(y, stack = "grp") \%>\%
  ebar(z, stack = "grp") \%>\%
  ebar(w, "grp2") \%>\%
  etheme("macarons") \%>\%
  etooltip(trigger = "axis")

}
\seealso{
\href{http://echarts.baidu.com/echarts2/doc/option-en.html#series-i(bar)}{official bar options docs}
}
