% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{ebar}
\alias{ebar}
\title{Add bars}
\usage{
ebar(p, serie, name = NULL, stack = NULL, clickable = TRUE,
  xAxisIndex = 0, yAxisIndex = 0, barGap = "100\%",
  barCategoryGap = "20\%", legendHoverLink = TRUE, z = 2, zlevel = 0,
  tooltip, itemStyle, barWidth, barMaxWidth, ...)
}
\description{
Add bar serie.
}
\examples{
mtcars \%>\%
  echart(mpg) \%>\%
  ebar(qsec)

mtcars \%>\%
  echart(disp) \%>\%
  ebar(mpg, stack = "grp") \%>\% # stack
  ebar(qsec, stack = "grp") \%>\% # stack
  ebar(wt) \%>\% # not stacked
  etooltip() \%>\%
  elegend()

}

