% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{ecloud}
\alias{ecloud}
\title{add wordcloud}
\usage{
ecloud(p, freq, color, name = NULL, clickable = TRUE,
  center = list("50\%", "50\%"), size = list("40\%", "40\%"),
  textRotation = list(0, 90), autoSize = list(enable = TRUE, minSize = 12),
  z = 2, zlevel = 0, markPoint, markLine, tooltip, ...)
}
\description{
add wordcloud
}
\examples{
tf <- data.frame(terms = c("ECharts", "htmlwidgets", "rstats", "htmltools"),
  freq = c(20, 17, 15, 7), color = c("Red", "orange", "yellow", "grey"))

tf \%>\%
  echart(terms) \%>\%
  ecloud(freq, color)

}

