% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{ecloud}
\alias{ecloud}
\alias{ecloud_}
\title{Add wordcloud}
\usage{
ecloud(p, freq, color, name = NULL, clickable = TRUE,
  center = list("50\%", "50\%"), size = list("40\%", "40\%"),
  textRotation = list(0, 90), autoSize = list(enable = TRUE, minSize = 12),
  z = 2, zlevel = 0, tooltip, ...)

ecloud_(p, freq, color = NULL, name = NULL, clickable = TRUE,
  center = list("50\%", "50\%"), size = list("40\%", "40\%"),
  textRotation = list(0, 90), autoSize = list(enable = TRUE, minSize = 12),
  z = 2, zlevel = 0, tooltip, ...)
}
\arguments{
\item{p}{an echart object.}

\item{freq}{frequencies.}

\item{color}{color of terms.}

\item{name}{name of wordcloud.}

\item{clickable}{whether terms are clickable.}

\item{center}{center of cloud.}

\item{size}{size of cloud.}

\item{textRotation}{horizontal and vertical text rotation.}

\item{autoSize}{automatic text size computation.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{tooltip}{cutomise tooltip.}

\item{...}{any other argument to pass to funnel.}
}
\description{
Add wordcloud serie.
}
\examples{
tf <- data.frame(terms = c("ECharts", "htmlwidgets", "rstats", "htmltools"),
  freq = c(20, 17, 15, 7), color = c("red", "orange", "yellow", "grey"))

tf \%>\%
  echart_("terms") \%>\%
  ecloud_("freq", "color") \%>\%
  etooltip()

}
\seealso{
\href{http://echarts.baidu.com/echarts2/doc/option-en.html#series-i(wordCloud)}{official wordcloud docs}
}

