% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.R, R/force_.R
\name{eforce}
\alias{eforce}
\alias{eforce_}
\title{Build force network}
\usage{
eforce(p, name = NULL, large = FALSE, center = list("50\%", "50\%"),
  roam = FALSE, size = "100\%", ribbonType = FALSE, minRadius = 10,
  maxRadius = 20, linkSymbol = "none", linkSymbolSize = list(10, 15),
  scaling = 1, gravity = 1, draggable = TRUE, useWorker = TRUE,
  steps = 1, z = 2, zlevel = 0, ...)

eforce_(p, name = NULL, large = FALSE, center = list("50\%", "50\%"),
  roam = FALSE, size = "100\%", ribbonType = FALSE, minRadius = 10,
  maxRadius = 20, linkSymbol = "none", linkSymbolSize = list(10, 15),
  scaling = 1, gravity = 1, draggable = TRUE, useWorker = TRUE,
  steps = 1, z = 2, zlevel = 0, ...)
}
\arguments{
\item{p}{an echart objects.}

\item{name}{name of network.}

\item{large}{set to \code{TRUE} to optimise for large graphs.}

\item{center}{center of network.}

\item{roam}{set to \code{TRUE} to enable zoom and drag.}

\item{size}{size of layout.}

\item{ribbonType}{whether to use ribbons.}

\item{minRadius, maxRadius}{minimum and maximum radius of nodes.}

\item{linkSymbol}{can be set to \code{arrow}.}

\item{linkSymbolSize}{size of \code{symbol}.}

\item{scaling}{scaling factor.}

\item{gravity}{centripetal force coefficient.}

\item{draggable}{set to \code{TRUE} to allow dragging nodes.}

\item{useWorker}{specifies whether to put layout calculation into web worker when the browser supports web worker.}

\item{steps}{the number of iterations of each frame layout calculation.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{...}{any other options to pass to serie.}
}
\description{
Build force network

Plot force directed graph.
}
\examples{
let <- LETTERS[1:20]

edges <- data.frame(source = sample(let, 20), target = sample(let, 20),
  weight = runif(20, 5, 20))

nodes <- data.frame(name = let, value = runif(20, 5, 25), group = rep(LETTERS[1:4], 5))

echart() \%>\%
  eforce(itemStyle = list(normal = list(label = list(show = TRUE)))) \%>\% # show labels
  enodes(nodes, name, value = value, category = group) \%>\%
  elinks(edges, source, target)

}
\seealso{
\code{\link{enodes}} \code{\link{eforce}}
}
