% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.R
\name{eforce}
\alias{eforce}
\title{Build force network}
\usage{
eforce(p, large = FALSE, center = list("50\%", "50\%"), roam = FALSE,
  size = "100\%", ribbonType = FALSE, minRadius = 10, maxRadius = 20,
  linkSymbol = "none", linkSymbolSize = list(10, 15), scaling = 1,
  gravity = 1, draggable = TRUE, useWorker = TRUE, steps = 1, z = 2,
  zlevel = 0, ...)
}
\description{
Plot force directed graph.
}

