% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{efunnel}
\alias{efunnel}
\alias{efunnel_}
\title{Add funnel}
\usage{
efunnel(p, serie, name = NULL, clickable = TRUE, legendHoverLink = TRUE,
  sort = "descending", min = NULL, max = NULL, x = 80, y = 60,
  x2 = 80, y2 = 60, width = NULL, height = NULL,
  funnelAlign = "center", minSize = "0\%", maxSize = "100\%", gap = 0,
  tooltip, ...)

efunnel_(p, serie, name = NULL, clickable = TRUE, legendHoverLink = TRUE,
  sort = "descending", min = 0, max = NULL, x = 80, y = 60, x2 = 80,
  y2 = 60, width = NULL, height = NULL, funnelAlign = "center",
  minSize = "0\%", maxSize = "100\%", gap = 0, tooltip, ...)
}
\arguments{
\item{p}{an echart object.}

\item{serie}{values to plot.}

\item{name}{name of serie.}

\item{clickable}{whether segments are clickable.}

\item{legendHoverLink}{enables legend hover link.}

\item{sort}{data sorting, takes \code{descending} or \code{ascending}.}

\item{min, max}{minimum and maximum values of funnel.}

\item{x, y, x2, y2}{coordinates of funnel.}

\item{width, height}{width and height of funnel.}

\item{funnelAlign}{alignment of funnel takes \code{left}, \code{right} and \code{center}.}

\item{minSize, maxSize}{minimum and maximum size of funnel.}

\item{gap}{gap between segments.}

\item{tooltip}{cutomise tooltip.}

\item{...}{any other argument to pass to funnel.}
}
\description{
Add funnel

Add funnel
}
\examples{
funnel <- data.frame(stage = c("View", "Click", "Purchase"), value = c(80, 30, 20))

funnel \%>\%
  echart(stage) \%>\%
  efunnel(value)

funnel <- data.frame(stage = c("View", "Click", "Purchase"), value = c(80, 30, 20))

funnel \%>\%
  echart_("stage") \%>\%
  efunnel_("value")

}
\seealso{
\href{http://echarts.baidu.com/echarts2/doc/option-en.html#series-i(funnel)}{official funnel docs}
}

