% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{efunnel}
\alias{efunnel}
\title{Add funnel}
\usage{
efunnel(p, serie, clickable = TRUE, legendHoverLink = TRUE,
  sort = "descending", min = 0, max = 100, x = 80, y = 60, x2 = 80,
  y2 = 60, width = NULL, height = NULL, funnelAlign = "center",
  minSize = "0\%", maxSize = "100\%", gap = 0, markPoint, markLine,
  tooltip, itemStyle, ...)
}
\description{
Add funnel
}
\examples{
funnel <- data.frame(stage = c("View", "Click", "Purchase"), value = c(80, 30, 20))

funnel \%>\%
  echart(stage) \%>\%
  efunnel(value)

}

