% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{egauge}
\alias{egauge}
\title{Add gauge}
\usage{
egauge(p, value, indicator = "", name = NULL, clickable = TRUE,
  legendHoverLink = TRUE, center = list("50\%", "50\%"),
  radius = list("0\%", "75\%"), startAngle = 225, endAngle = -45,
  min = 0, max = 100, splitNumber = 10, z = 2, zlevel = 0, tooltip,
  markPoint, markLine, axisLine, axisLabel, splitLine, pointer, title, detail,
  ...)
}
\description{
Add gauge
}
\examples{
echart() \%>\%
  egauge(85, "SPEED")

}

