% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{eheatmap}
\alias{eheatmap}
\title{add heatmap}
\usage{
eheatmap(p, y, z, ...)
}
\description{
add heatmap
}
\examples{
set.seed(19880525)
matrix <- data.frame(x = runif(100, 10, 200), y = runif(100, 10, 200), z = runif(100, 10 , 200))

matrix \%>\%
  echart(x) \%>\%
  eheatmap(y, z)

}

