% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{eheatmap}
\alias{eheatmap}
\title{add heatmap}
\usage{
eheatmap(p, y, values, name = NULL, clickable = TRUE, blurSize = 30,
  minAlpha = 0.5, valueScale = 1, opacity = 1, z = 2, zlevel = 0,
  gradientColors, markPoint, markLine, tooltip, ...)
}
\description{
add heatmap
}
\examples{
set.seed(19880525)
matrix <- data.frame(x = runif(100, 10, 200), y = runif(100, 10, 200), z = runif(100, 10 , 200))

matrix \%>\%
  echart(x) \%>\%
  eheatmap(y, z)

}

