% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{eline}
\alias{eline}
\title{Add lines}
\usage{
eline(p, serie, name = NULL, stack = NULL, clickable = TRUE,
  xAxisIndex = 0, yAxisIndex = 0, symbol = NULL, symbolSize = "2 | 4",
  symbolRate = NULL, showAllSymbol = FALSE, smooth = TRUE,
  legendHoverLink = TRUE, dataFilter = "nearest", z = 2, zlevel = 0,
  tooltip, markPoint, markLine, ...)
}
\description{
Add line serie.
}
\examples{
mtcars \%>\%
  echart(mpg) \%>\%
  eline(qsec)

mtcars \%>\%
  echart(disp) \%>\%
  eline(mpg, stack = "grp") \%>\%
  eline(qsec, stack = "grp") \%>\%
  eline(wt, fill = TRUE) \%>\%
  etooltip()

}

