% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{eline}
\alias{eline}
\alias{eline_}
\title{Add lines}
\usage{
eline(p, serie, name = NULL, stack = NULL, clickable = TRUE,
  xAxisIndex = 0, yAxisIndex = 0, symbol = NULL, symbolSize = "2 | 4",
  symbolRotate = NULL, showAllSymbol = FALSE, smooth = TRUE,
  legendHoverLink = TRUE, dataFilter = "nearest", z = 2, zlevel = 0,
  tooltip, ...)

eline_(p, serie, name = NULL, stack = NULL, clickable = TRUE,
  xAxisIndex = 0, yAxisIndex = 0, symbol = NULL, symbolSize = "4",
  symbolRotate = NULL, showAllSymbol = FALSE, smooth = TRUE,
  legendHoverLink = TRUE, dataFilter = "nearest", z = 2, zlevel = 0,
  tooltip, ...)
}
\arguments{
\item{p}{an echart object.}

\item{serie}{value column name to plot.}

\item{name}{of serie.}

\item{stack}{name of the stack.}

\item{clickable}{whether plot is clickable.}

\item{xAxisIndex, yAxisIndex}{axis indexes.}

\item{symbol}{symbol for point marker, see details for valid values.}

\item{symbolSize}{of symbol.}

\item{symbolRotate}{angle by which symbol is rotated, i.e.: \code{30}.}

\item{showAllSymbol}{By default, a symbol will show only when its corresponding axis label does.}

\item{smooth}{whether to smooth line.}

\item{legendHoverLink}{enables legend hover link to the chart.}

\item{dataFilter}{ECharts data filtering strategy, see details.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{tooltip}{style of tooltip.}

\item{...}{any other argument to pass to the serie.}
}
\description{
Add lines

Add line serie.
}
\details{
Valid values for \code{symbol}:
\itemize{
  \item{\code{circle}}
  \item{\code{rectangle}}
  \item{\code{triangle}}
  \item{\code{diamond}}
  \item{\code{emptyCircle}}
  \item{\code{emptyRectangle}}
  \item{\code{emptyTriangle}}
  \item{\code{emptyDiamond}}
  \item{\code{heart}}
  \item{\code{droplet}}
  \item{\code{pin}}
  \item{\code{arrow}}
  \item{\code{star}}
}

\code{dataFilter}: ECharts will optimize for the situation when data number is much larger than viewport width.
It will filter the data showed in one pixel width. And this option is for data filtering strategy.

Valid values for \code{dataFilter} are:
\itemize{
  \item{\code{nearest} (default)}
  \item{\code{min}}
  \item{\code{max}}
  \item{\code{average}}
}
}
\examples{
mtcars \%>\%
  echart(mpg) \%>\%
  eline(qsec)

mtcars \%>\%
  echart(disp) \%>\%
  eline(mpg, stack = "grp") \%>\%
  eline(qsec, stack = "grp") \%>\%
  eline(wt, fill = TRUE) \%>\%
  etooltip()

mtcars \%>\%
  echart_("mpg") \%>\%
  eline_("qsec", symbol = "emptyCircle") \%>\%
  etooltip(trigger = "item")

mtcars \%>\%
  echart_("disp") \%>\%
  eline_("mpg", stack = "grp") \%>\%
  eline_("qsec", stack = "grp") \%>\%
  eline_("wt", smooth = FALSE) \%>\%
  etooltip() \%>\%
  elegend() \%>\%
  etoolbox_magic(type = list("line", "bar"))

}
\seealso{
\href{http://echarts.baidu.com/echarts2/doc/option-en.html#series-i(line)}{official line options docs}
}

