% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{emap}
\alias{emap}
\alias{emap_}
\title{Add blank map}
\usage{
emap(p, name = NULL, mapType = "world", clickable = TRUE, z = 2,
  zlevel = 0, selectedMode = NULL, hoverable = FALSE,
  dataRangeHoverLink = TRUE, mapLocation = list(x = "center", y = "center"),
  mapValueCalculation = "sum", mapValuePrecision = 0,
  showLegendSymbol = TRUE, roam = FALSE, scaleLimit = NULL,
  nameMap = NULL, textFixed = NULL, ...)

emap_(p, name = NULL, mapType = "world", clickable = TRUE, z = 2,
  zlevel = 0, selectedMode = NULL, hoverable = FALSE,
  dataRangeHoverLink = TRUE, mapLocation = list(x = "center", y = "center"),
  mapValueCalculation = "sum", mapValuePrecision = 0,
  showLegendSymbol = TRUE, roam = FALSE, scaleLimit = NULL,
  nameMap = NULL, textFixed = NULL, ...)
}
\arguments{
\item{p}{an echart object.}

\item{name}{name of serie.}

\item{mapType}{type of map, see examples.}

\item{clickable}{whether elements are clickable.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{selectedMode}{whether items can be selected.}

\item{hoverable}{whether elements are hoverable.}

\item{dataRangeHoverLink}{enables dataRange hover link to the chart.}

\item{mapLocation}{x and y location of map on canvas, takes \code{top}, \code{bottom}, \code{left}, \code{right}, \code{center}.}

\item{mapValueCalculation}{takes \code{sum} or \code{average}.}

\item{mapValuePrecision}{decimal precision.}

\item{showLegendSymbol}{whether to show symbol on legend.}

\item{roam}{enables zoom and drag.}

\item{scaleLimit}{controls drag and zoom limits.}

\item{nameMap}{custom name mapping.}

\item{textFixed}{fixed text location for a region.}

\item{...}{any other options to pass to map serie.}
}
\description{
Setup map plot.
}
\examples{
coords <- data.frame(city = c("London", "New York", "Beijing", "Sydney"),
  lon = c(-0.1167218, -73.98002, 116.3883, 151.18518),
  lat = c(51.49999, 40.74998, 39.92889, -33.92001),
  values = runif(4, 10, 20))

coords \%>\%
  echart_("city") \%>\% # initialise chart
  emap_() \%>\% # setup default map
  emap_coords_("lon", "lat") \%>\% # add coordinates
  emap_points_("values") # plot values on coordinates

edges <- data.frame(source = c("Beijing", "Beijing", "New York"),
  target = c("Sydney", "London", "London"))

coords \%>\%
  echart_("city") \%>\%
  emap_() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_lines_(edges, "source", "target")

data <- data.frame(lon = runif(200, 90, 120),
  lat = runif(200, 30, 39),
  z = runif(200, 50, 75))

data \%>\%
  echart_() \%>\%
  emap_(mapType = "china") \%>\%
  emap_heat_("lon", "lat", "z")

us_data <- data.frame(state = c("New York", "Los Angeles", "Dallas"),
                      lat = c(40.730610, 34.052235, 33.940369),
                      lon = c(-73.935242, -118.243683, -84.692894),
                      values = round(runif(3, 1, 2)))

us_data \%>\%
  echart_("state") \%>\%
  emap(mapType = "world|United States of America") \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_points_("values")

}
\seealso{
\code{\link{emap_coords}}, \code{\link{emap_heat}}, \code{\link{emap_lines}}, \code{emap_choropleth},
\code{\link{emap_points}}, \href{http://echarts.baidu.com/echarts2/doc/option-en.html#series-i(map)}{official map docs}
}

