% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{emap_choropleth}
\alias{emap_choropleth}
\alias{emap_choropleth}
\alias{emap_choropleth_}
\title{Add choropleth}
\usage{
emap_choropleth(p, serie)

emap_choropleth_(p, serie)
}
\arguments{
\item{p}{an echart object.}

\item{serie}{values to plot.}
}
\description{
Add choropleth
}
\examples{
choropleth <- data.frame(countries = c("France", "Brazil", "China", "Russia", "Canada", "India"),
  values = round(runif(6, 10, 25)))

choropleth \%>\%
  echart_("countries") \%>\%
  emap_() \%>\%
  emap_choropleth_("values")

choropleth \%>\%
  echart_("countries") \%>\%
  emap() \%>\%
  emap_choropleth(values) \%>\%
  ecolorbar(color = list("red", "yellow"), calculable = TRUE)

}
\seealso{
\code{\link{ecolorbar}}
}
