% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{emap_heat}
\alias{emap_heat}
\title{Add heat on map}
\usage{
emap_heat(p, lat, z, mapType = "world", ...)
}
\description{
Add heat on map
}
\examples{
data <- data.frame(lon = runif(200, 90, 120),
  lat = runif(200, 30, 39),
  z = runif(200, 50, 75))

data \%>\%
  echart(lon) \%>\%
  emap_heat(lat, z, "china")

}

