% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{emap_heat_}
\alias{emap_heat_}
\title{Add heat on map}
\usage{
emap_heat_(p, lon, lat, z, blurSize = 30, minAlpha = 0.05, valueScale = 1,
  opacity = 1, gradientColors, ...)
}
\description{
Add heat on map
}
\examples{
data <- data.frame(lon = runif(200, 90, 120),
  lat = runif(200, 30, 39),
  z = runif(200, 50, 75))

data \%>\%
  echart_() \%>\%
  emap(mapType = "china") \%>\%
  emap_heat_("lon", "lat", "z")

data \%>\%
  echart() \%>\%
  emap(mapType = "china") \%>\%
  emap_heat_("lon", "lat", "z", blurSize = 50, minAlpha = 0.3, opacity = 0.8) \%>\%
  etheme("dark")

}

