% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{emap_line_effect}
\alias{emap_line_effect}
\title{emap line effect}
\usage{
emap_line_effect(show = TRUE, loop = TRUE, period = 30, scaleSize = 1,
  color = "#fff", shadowBlur = 10, shadowColor = NULL)
}
\description{
Effect for emap lines
}

