% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{emap_lines}
\alias{emap_lines}
\alias{emap_lines}
\alias{emap_lines_}
\title{Add map lines}
\usage{
emap_lines(p, edges, source, target, name = NULL, clickable = TRUE,
  symbol = "arrow", symbolSize = 2, symbolRotate = NULL, large = FALSE,
  smooth = TRUE, z = 2, zlevel = 0, smoothness = 0.2, precision = 2,
  bundling = list(enable = FALSE, maxTurningAngle = 45), ...)

emap_lines_(p, edges, source, target, name = NULL, clickable = TRUE,
  symbol = "arrow", symbolSize = 2, symbolRotate = NULL, large = FALSE,
  smooth = TRUE, z = 2, zlevel = 0, smoothness = 0.2, precision = 2,
  bundling = list(enable = FALSE, maxTurningAngle = 45), ...)
}
\arguments{
\item{p}{an echart object.}

\item{edges}{edges data.frame.}

\item{source, target}{source and target columns in edges data.frame.}

\item{name}{name of serie.}

\item{clickable}{whether lines are clikable.}

\item{symbol}{symbol, see valid details for valid values.}

\item{symbolSize}{of symbol.}

\item{symbolRotate}{angle by which symbol is rotated, i.e.: \code{30}.}

\item{large}{optimises for 2'000 data points and over.}

\item{smooth}{whether to smooth lines.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{smoothness}{line smoothness}

\item{precision}{for 'average'.}

\item{bundling}{edge bundling settings, see usage.}

\item{...}{any other options to pass to line.}
}
\description{
Add lines on map.
}
\details{
Valid values for \code{symbol}:
\itemize{
  \item{\code{circle}}
  \item{\code{rectangle}}
  \item{\code{triangle}}
  \item{\code{diamond}}
  \item{\code{emptyCircle}}
  \item{\code{emptyRectangle}}
  \item{\code{emptyTriangle}}
  \item{\code{emptyDiamond}}
  \item{\code{heart}}
  \item{\code{droplet}}
  \item{\code{pin}}
  \item{\code{arrow}}
  \item{\code{star}}
}
}
\examples{
coords <- data.frame(city = c("London", "New York", "Beijing", "Sydney"),
  lon = c(-0.1167218, -73.98002, 116.3883, 151.18518),
  lat = c(51.49999, 40.74998, 39.92889, -33.92001))

edges <- data.frame(source = c("Beijing", "Beijing", "New York"),
  target = c("Sydney", "London", "London"))

coords \%>\%
  echart_("city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_lines_(edges, "source", "target")

edges2 <- data.frame(source = "London", target = "Sydney")

coords \%>\%
  echart_("city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_lines_(edges, "source", "target") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_lines_(edges2, "source", "target", effect = emap_line_effect()) \%>\%
  etheme("helianthus")

coords2 <- data.frame(city = "Sydney", lon = 151.18518, lat = -33.92001, value = 20)

coords \%>\%
  echart_("city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_lines_(edges, "source", "target") \%>\%
  edata_(coords2, "city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_lines_(edges2, "source", "target", effect = emap_line_effect(scaleSize = 2)) \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_points_("value", symbol = "emptyCircle", effect = list(show = TRUE, shadowBlur = 10)) \%>\%
  etheme("dark")

}
\seealso{
\code{\link{emap_coords}} \href{http://echarts.baidu.com/echarts2/doc/option-en.html#series-i(map).markLine}{official map line docs}
}
