% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{emap_points}
\alias{emap_points}
\alias{emap_points_}
\title{Add map points}
\usage{
emap_points(p, serie, clickable = TRUE, symbol = "pin", symbolSize = 10,
  symbolRotate = NULL, large = FALSE, itemStyle = NULL, ...)

emap_points_(p, serie, clickable = TRUE, symbol = "pin", symbolSize = 10,
  symbolRotate = NULL, large = FALSE, itemStyle = NULL, ...)
}
\arguments{
\item{p}{an echart objects.}

\item{serie}{values to plot.}

\item{clickable}{whether points are clickable.}

\item{symbol}{point symbol, see details for valid values.}

\item{symbolSize}{size of points.}

\item{symbolRotate}{angle by which symbol is rotated, i.e.: \code{30}.}

\item{large}{whether to optimise for large datasets: 2K points +.}

\item{itemStyle}{cutomise points.}

\item{...}{any other option to pass to points.}
}
\description{
Add map points
}
\details{
Valid values for \code{symbol}:
\itemize{
  \item{\code{circle}}
  \item{\code{rectangle}}
  \item{\code{triangle}}
  \item{\code{diamond}}
  \item{\code{emptyCircle}}
  \item{\code{emptyRectangle}}
  \item{\code{emptyTriangle}}
  \item{\code{emptyDiamond}}
  \item{\code{heart}}
  \item{\code{droplet}}
  \item{\code{pin}}
  \item{\code{arrow}}
  \item{\code{star}}
}
}
\examples{
coords <- data.frame(city = c("London", "New York", "Beijing", "Sydney"),
  lon = c(-0.1167218, -73.98002, 116.3883, 151.18518),
  lat = c(51.49999, 40.74998, 39.92889, -33.92001),
  values = runif(4, 10, 20))

coords \%>\%
  echart_("city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_points_("values")

coords2 <- data.frame(city = "Rio", lon = -43.172896, lat = -22.906847, value = 15)

coords \%>\%
  echart_("city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_points_("values", symbolSize = 5) \%>\%
  edata_(coords2, "city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_points_("value", symbol = "emptyCircle", effect = list(show = TRUE, shadowBlur = 10)) \%>\%
  etheme("helianthus")

}
\seealso{
\href{http://echarts.baidu.com/echarts2/doc/option-en.html#series-i(map).markPoint}{office map points docs}
}

