% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{emap_points_}
\alias{emap_points_}
\title{Add map points}
\usage{
emap_points_(p, serie, clickable = TRUE, symbol = "pin", symbolSize = 10,
  symbolRotate = NULL, large = FALSE, itemStyle = NULL, ...)
}
\description{
Add map points
}
\examples{
coords <- data.frame(city = c("London", "New York", "Beijing", "Sydney"),
  lon = c(-0.1167218, -73.98002, 116.3883, 151.18518),
  lat = c(51.49999, 40.74998, 39.92889, -33.92001),
  values = runif(4, 10, 20))

coords \%>\%
  echart_("city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_points_("values")

coords2 <- data.frame(city = "Rio", lon = -43.172896, lat = -22.906847, value = 15)

coords \%>\%
  echart_("city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_points_("values", symbolSize = 5) \%>\%
  edata_(coords2, "city") \%>\%
  emap() \%>\%
  emap_coords_("lon", "lat") \%>\%
  emap_points_("value", symbol = "emptyCircle", effect = list(show = TRUE, shadowBlur = 10)) \%>\%
  etheme("helianthus")

}

