% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.R, R/force_.R
\name{nodes}
\alias{nodes}
\alias{enodes}
\alias{enodes_}
\title{Add nodes}
\usage{
enodes(p, nodes, name, label, value, category, symbolSize, depth,
  ignore = FALSE, symbol = "circle", fixX = FALSE, fixY = FALSE)

enodes_(p, nodes, name, label = NULL, value = NULL, category = NULL,
  symbolSize = NULL, depth = NULL, ignore = FALSE, symbol = "circle",
  fixX = FALSE, fixY = FALSE)
}
\arguments{
\item{p}{an echart object.}

\item{nodes}{nodes data.frame.}

\item{name}{name column.}

\item{label}{nodes label column.}

\item{value}{nodes value (size).}

\item{category}{nodes group column.}

\item{symbolSize}{nodes symbol size column.}

\item{depth}{depth of nodes.}

\item{ignore}{whether to ignore nodes.}

\item{symbol}{nodes symbol, see details for valid values.}

\item{fixX, fixY}{whether to fix x and y axis position.}
}
\description{
Add nodes for \code{\link{eforce}}.

Add nodes for \code{\link{eforce}}.
}
\details{
Valid values for \code{symbol}:
\itemize{
  \item{\code{circle}}
  \item{\code{rectangle}}
  \item{\code{triangle}}
  \item{\code{diamond}}
  \item{\code{emptyCircle}}
  \item{\code{emptyRectangle}}
  \item{\code{emptyTriangle}}
  \item{\code{emptyDiamond}}
  \item{\code{heart}}
  \item{\code{droplet}}
  \item{\code{pin}}
  \item{\code{arrow}}
  \item{\code{star}}
}
}
\examples{
let <- LETTERS[1:20]

edges <- data.frame(source = sample(let, 20), target = sample(let, 20),
  weight = runif(20, 5, 20))

nodes <- data.frame(name = let, value = runif(20, 5, 25), group = rep(LETTERS[1:4], 5))

echart() \%>\%
  eforce(itemStyle = list(normal = list(label = list(show = TRUE)))) \%>\% # show labels
  enodes(nodes, name, value = value, category = group) \%>\%
  elinks(edges, source, target)

let <- LETTERS[1:20]

edges <- data.frame(source = sample(let, 20), target = sample(let, 20),
  weight = runif(20, 5, 20))

nodes <- data.frame(name = let, value = runif(20, 5, 25), group = rep(LETTERS[1:4], 5))

echart() \%>\%
  eforce_(itemStyle = list(normal = list(label = list(show = TRUE)))) \%>\% # show labels
  enodes_(nodes, "name", value = "value", category = "group") \%>\%
  elinks_(edges, "source", "target")

}
\seealso{
\code{\link{enodes}} \code{\link{eforce}}
}
