% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{eoptions}
\alias{eoptions}
\title{Add global options}
\usage{
eoptions(p, backgroundColor = NULL, renderAsImage = FALSE,
  calculable = FALSE, color = NULL, symbolList = NULL, ...)
}
\arguments{
\item{p}{an echart object.}

\item{backgroundColor}{background color.}

\item{renderAsImage}{allows rendering as image.}

\item{calculable}{specifies whether the drag-recalculate feature will be enabled.}

\item{color}{colors to use in chart.}

\item{symbolList}{list of default symbols.}

\item{...}{any other options.}
}
\description{
Add global options.
}
\examples{
mtcars \%>\%
  echart(mpg) \%>\%
  eline(qsec) \%>\%
  eoptions(backgroundColor = "black")

}
