% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{eoptions}
\alias{eoptions}
\title{Add global options}
\usage{
eoptions(p, backgroundColor = NULL, color, renderAsImage = FALSE,
  calculable = FALSE, symbolList, calculableColor = "rgba(255,165,0,0.6)",
  calculableHolderColor = "#ccc", nameConnector = "&",
  valueConnector = ":", ...)
}
\description{
Add global options
}

