% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{epie}
\alias{epie}
\title{Add pie}
\usage{
epie(p, serie, name = NULL, clickable = TRUE, legendHoverLink = TRUE,
  center = list("50\%", "50\%"), radius = list(0, "75\%"),
  startAngle = 90, minAngle = 0, clockWise = TRUE, roseType = NULL,
  selectedOffset = 10, selectedMode = TRUE, z = 2, zlevel = 0, ...)
}
\description{
Add pie chart
}
\examples{
pie <- data.frame(name = c("banana", "apple", "pineapple", "onion"),
  value = c(40, 15, 12, 9))

pie \%>\%
  echart(name) \%>\%
  epie(value)

pie \%>\%
  echart(name) \%>\%
  epie(value, roseType = "area") \%>\%
  etheme("helianthus")

pie \%>\%
  echart(name) \%>\%
  epie(value, roseType = "radius") \%>\%
  etheme("mint")

}

