% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{epie}
\alias{epie}
\alias{epie_}
\title{Add pie}
\usage{
epie(p, serie, name = NULL, clickable = TRUE, legendHoverLink = TRUE,
  center = list("50\%", "50\%"), radius = list(0, "75\%"),
  startAngle = 90, minAngle = 0, clockWise = TRUE, roseType = NULL,
  selectedOffset = 10, selectedMode = TRUE, z = 2, zlevel = 0, ...)

epie_(p, serie, name = NULL, clickable = TRUE, legendHoverLink = TRUE,
  center = list("50\%", "50\%"), radius = list(0, "75\%"),
  startAngle = 90, minAngle = 0, clockWise = TRUE, roseType = NULL,
  selectedOffset = 10, selectedMode = TRUE, z = 2, zlevel = 0, ...)
}
\arguments{
\item{p}{an echart object.}

\item{serie}{value column name to plot.}

\item{name}{of serie.}

\item{clickable}{whether plot is clickable.}

\item{legendHoverLink}{enables legend hover links.}

\item{center}{coordinates of the center.}

\item{radius}{radius in pixels or percent.}

\item{startAngle, minAngle}{start and minimum angle.}

\item{clockWise}{whether to display slices in clockwise direction}

\item{roseType}{type of pie, takes \code{NULL}, \code{area} or \code{radius}, see examples.}

\item{selectedOffset}{offset of selected slice.}

\item{selectedMode}{whether slices are selectable.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{...}{any other option to pass to serie.}
}
\description{
Add pie

Add pie chart
}
\examples{
pie <- data.frame(name = c("banana", "apple", "pineapple", "onion"),
  value = c(40, 15, 12, 9))

pie \%>\%
  echart(name) \%>\%
  epie(value)

pie \%>\%
  echart(name) \%>\%
  epie(value, roseType = "area") \%>\%
  etheme("helianthus")

pie \%>\%
  echart(name) \%>\%
  epie(value, roseType = "radius") \%>\%
  etheme("mint")

pie <- data.frame(name = c("banana", "apple", "pineapple", "onion"),
  value = c(26, 15, 12, 9))

pie \%>\%
  echart_("name") \%>\%
  epie_("value")

pie \%>\%
  echart_("name") \%>\%
  epie_("value", roseType = "area") \%>\%
  etheme("helianthus")

pie \%>\%
  echart_("name") \%>\%
  epie_("value", roseType = "radius") \%>\%
  etheme("mint")

}
\seealso{
\href{http://echarts.baidu.com/echarts2/doc/option-en.html#series-i(pie)}{official pie options docs}
}

