% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{eradar_}
\alias{eradar_}
\title{add radar}
\usage{
eradar_(p, serie, name = NULL, ...)
}
\description{
add radar
}
\examples{
radar <- data.frame(axis = LETTERS[1:6], value = runif(6, 2, 10), value2 = runif(6, 3, 11))

radar \%>\%
  echart_("axis") \%>\%
  eradar_("value") \%>\%
  eradar_("value2") \%>\%
  elegend()

}

