% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{eradar_}
\alias{eradar_}
\title{Add radar}
\usage{
eradar_(p, serie, name = NULL, clickable = TRUE, symbol = NULL,
  symbolSize = 4, symbolRotate = NULL, legendHoverLink = TRUE,
  polarIndex = 0, z = 2, zlevel = 0, ...)
}
\arguments{
\item{p}{an echart object.}

\item{serie}{value column name to plot.}

\item{name}{of serie.}

\item{clickable}{whether plot is clickable.}

\item{symbol}{marker, see details for valid values.}

\item{symbolSize}{of symbol.}

\item{symbolRotate}{angle by which symbol is rotated, i.e.: \code{30}.}

\item{legendHoverLink}{enables legend hover links.}

\item{polarIndex}{polar coordinates index.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{...}{any other options to pass to the serie.}
}
\description{
Add radar chart.
}
\details{
Valid values for \code{symbol}:
\itemize{
  \item{\code{circle}}
  \item{\code{rectangle}}
  \item{\code{triangle}}
  \item{\code{diamond}}
  \item{\code{emptyCircle}}
  \item{\code{emptyRectangle}}
  \item{\code{emptyTriangle}}
  \item{\code{emptyDiamond}}
  \item{\code{heart}}
  \item{\code{droplet}}
  \item{\code{pin}}
  \item{\code{arrow}}
  \item{\code{star}}
}
}
\examples{
radar <- data.frame(axis = rep(LETTERS[1:6], 4), grp = rep(LETTERS[4:9], 4),
  value = runif(24, 2, 10))

radar \%>\%
  group_by_("grp") \%>\%
  echart_("axis") \%>\%
  eradar_("value") \%>\%
  elegend()

}
\seealso{
\href{http://echarts.baidu.com/echarts2/doc/option-en.html#series-i(radar)}{official radar options docs}
}

