% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{escatter}
\alias{escatter}
\title{Add scatter}
\usage{
escatter(p, serie, size, ...)
}
\description{
Add scatter serie.
}
\examples{
mtcars \%>\%
  echart(disp) \%>\%
  escatter(mpg, qsec * 3)

mtcars \%>\%
  echart(disp) \%>\%
  escatter(mpg, disp / 8) \%>\%
  escatter(qsec, disp / 8)

}

