% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{escatter}
\alias{escatter}
\title{Add scatter}
\usage{
escatter(p, serie, size, name = NULL, clickable = TRUE, ...)
}
\description{
Add scatter serie.
}
\examples{
mtcars \%>\%
  echart(disp) \%>\%
  escatter(mpg)

mtcars \%>\%
  echart(disp) \%>\%
  escatter(mpg, mpg * 2) \%>\%
  escatter(qsec, qsec * 2)

}

