% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{escatter_}
\alias{escatter_}
\title{Add scatter}
\usage{
escatter_(p, serie, size, name = NULL, clickable = TRUE, ...)
}
\description{
Add scatter serie.
}
\examples{
mtcars \%>\%
  echart_("disp") \%>\%
  escatter_("mpg")

mtcars \%>\%
  echart_("disp") \%>\%
  escatter_("mpg", "mpg") \%>\%
  escatter_("qsec", "qsec")

}

