% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{etimeline_}
\alias{etimeline_}
\title{Add timeline}
\usage{
etimeline_(p, timeline, show = TRUE, zlevel = 0, z = 4, type = "time",
  notMerge = FALSE, realtime = TRUE, x = 80, y = NULL, x2 = 80,
  y2 = 0, height = 50, backgroundColor = "rgba(0,0,0,0)",
  borderWidth = 0, borderColor = "#ccc", padding = 5,
  controlPosition = "left", autoPlay = FALSE, loop = TRUE,
  playInterval = 2000, symbol = "emptyDiamond", symbolSize = 4,
  currentIndex = 0, ...)
}
\description{
Add timeline
}
\examples{
tl <- data.frame(x = 1:100,
  values = runif(100, 1, 5),
  time.stamp = rep(c("day1", "day2", "day3", "day4"), 25))

tl \%>\%
  echart_("x") \%>\%
  eline_("values") \%>\%
  etimeline_()

}

