% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{etoolbox}
\alias{etoolbox}
\title{Add toolbox}
\usage{
etoolbox(p, show = TRUE, zlevel = 0, z = 6, orient = "horizontal",
  x = "right", y = "top", backgroundColor = "rgba(0,0,0,0)",
  borderColor = "#ccc", borderWidth = 0, padding = 5, itemGap = 10,
  itemSize = 16, color, disableColor = "#ddd", effectiveColor = "red",
  showTitle = TRUE, textStyle, ...)
}
\description{
Add toolbox
}

