% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{etoolbox_magic}
\alias{etoolbox_magic}
\title{Add toolbox magic buttons}
\usage{
etoolbox_magic(p, show = TRUE, type = list(), title, ...)
}
\arguments{
\item{p}{an echart object.}

\item{show}{wehtehr to show magic buttons.}

\item{type}{chart types, see details.}

\item{title}{titles of charts.}

\item{...}{any other options to pass to magic feature.}
}
\description{
Enable changing chart type.
}
\details{
Pass a \code{list} to \code{type}, valid values are:

\itemize{
  \item{\code{line}}
  \item{\code{bar}}
  \item{\code{stack}}
  \item{\code{tiled}}
  \item{\code{force}}
  \item{\code{chord}}
  \item{\code{pie}}
  \item{\code{funnel}}
}
}
\examples{
mtcars \%>\%
  echart(disp) \%>\%
  ebar(mpg, stack = "grp") \%>\% # stack
  ebar(qsec, stack = "grp") \%>\% # stack
  ebar(wt) \%>\% # not stacked
  etooltip() \%>\%
  elegend() \%>\%
  etoolbox() \%>\%
  etoolbox_magic(type = list("bar", "line", "stack", "tiled"))

}

