% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{etoolbox_mark}
\alias{etoolbox_mark}
\title{Add toolbox feature mark button}
\usage{
etoolbox_mark(p, show = TRUE, title = list(mark = "Mark", markUndo = "Undo",
  markClear = "Clear"), lineStyle = list(color = "#1e90ff", typed = "dashed",
  width = 2, shadowColor = "rgba(0,0,0,0)", shadowBlur = 5, shadowOffsetX = 3,
  shadowOffsetY = 3))
}
\arguments{
\item{p}{an echart object.}

\item{show}{whether to show mark.}

\item{title}{mark button title.}

\item{lineStyle}{style of marked line.}
}
\description{
Enable marking chart.
}
\examples{
mtcars \%>\%
  echart(qsec) \%>\%
  ebar(mpg) \%>\%
  etoolbox() \%>\%
  etoolbox_mark()

}

