% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{etoolbox_save}
\alias{etoolbox_save}
\title{Add toolbox save as image button}
\usage{
etoolbox_save(p, show = TRUE, title = "Save as image", type = "png",
  name = "echarts", lang = "Save")
}
\description{
Add toolbox save as image button
}

