% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{etoolbox_zoom}
\alias{etoolbox_zoom}
\title{Add toolbox zoom button}
\usage{
etoolbox_zoom(p, show = TRUE, title = list(dataZoom = "Area Zoom",
  dataZoomReset = "Reset"))
}
\description{
Add toolbox zoom button
}

