% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/add_.R
\name{evenn}
\alias{evenn}
\alias{evenn_}
\title{Add venn}
\usage{
evenn(p, serie, name = NULL, clickable = TRUE, z = 2, zlevel = 0,
  tooltip = NULL, ...)

evenn_(p, serie, name = NULL, clickable = TRUE, z = 2, zlevel = 0,
  tooltip = NULL, ...)
}
\arguments{
\item{p}{an echart object.}

\item{serie}{a named vector, see details.}

\item{name}{name of serie.}

\item{clickable}{whether ciorcles are clickable.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{tooltip}{cutomise tooltip.}

\item{...}{any other argument to pass to funnel.}
}
\description{
Add venn

Add venn diagram
}
\examples{
venn <- data.frame(name = c("banana", "pineapple", "overlap"),
  values = c(20, 50, 10))

venn \%>\%
  echart(name) \%>\%
  evenn(values) \%>\%
  etheme("infographic")

venn <- data.frame(name = c("banana", "pineapple", "overlap"),
  values = c(20, 50, 10))

venn \%>\%
  echart_("name") \%>\%
  evenn_("values") \%>\%
  etheme("macarons2")

}
\seealso{
\href{http://echarts.baidu.com/echarts2/doc/option-en.html#series-i(venn)}{official venn docs}
}

