% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{evenn_}
\alias{evenn_}
\title{Add venn}
\usage{
evenn_(p, serie, name = NULL, clickable = TRUE, z = 2, zlevel = 0,
  tooltip, ...)
}
\arguments{
\item{serie}{a named vector, see details.}
}
\description{
Add venn diagram
}
\examples{
venn <- data.frame(name = c("banana", "pineapple", "overlap"),
  values = c(20, 50, 10))

venn \%>\%
  echart_("name") \%>\%
  evenn_("values") \%>\%
  etheme("mint")

}

