% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ezoom}
\alias{ezoom}
\title{Add data zoom}
\usage{
ezoom(p, show = TRUE, zlevel = 0, z = 4, orient = "horizontal",
  backgroundColor = "rgba(0,0,0,0)", dataBackgroundColor = "#eee",
  fillerColor = "rgba(144,197,237,0.2)",
  handleColor = "rgba(70,130,180,0.8)", handleSize = 8, start = 0,
  end = 100, showDetail = TRUE, realtime = FALSE, zoomLock = FALSE, ...)
}
\description{
Add data zoom
}

