% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{eget_options_p}
\alias{eget_options_p}
\alias{eadd_linedata_p}
\alias{etheme_p}
\title{echarts Shiny proxies}
\usage{
eget_options_p(proxy)

eadd_linedata_p(proxy, data, serie, head = FALSE, grow = FALSE)

etheme_p(proxy, theme)
}
\arguments{
\item{proxy}{as returned by \code{\link{echartsProxy}}}
}
\description{
Shiny proxies for echarts.
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage(
  titlePanel("echarts Proxies"),
  fluidRow(
    column(
      4,
      actionButton("getOptions", "get options")
    ),
    column(
      4,
      selectInput(
        "theme",
        "change theme",
        choices =
          c(
            "default",
            "red",
            "mint",
            "macarons",
            "macarons2",
            "green",
            "blue",
            "dark",
            "gray",
            "helianthus",
            "wef",
            "roma",
            "sakura",
            "shine",
            "infographic",
            "solarlight"
         )
      )
    ),
    column(
      4,
      actionButton("addData", "add data")
    )
  ),
  fluidRow(
    echartsOutput("chart")
  )
)

server <- function(input, output){
  df <- data.frame(x = 1:50, y = runif(50, 5, 10), z = runif(50, 7, 12), w = runif(50, 10, 13))

  output$chart <- renderEcharts(
    df \%>\%
      echart(x) \%>\%
      eline(y, name = "Metric", stack = "grp1") \%>\%
      earea(w, name = "Variable", stack = "grp2") \%>\%
      earea(z, name = "Variable", stack = "grp2") \%>\%
      etitle("echarts", "and its Shiny proxies")
  )

  reactive_data <- reactive({
    data.frame(x = 1:50, y = runif(50, 5, 10), z = runif(50, 7, 12), w = runif(50, 10, 13))
  })

  observeEvent(input$getOptions, {
    echartsProxy("chart") \%>\%
      egetoptions_p()
  })

  observeEvent(input$theme, {
    echartsProxy("chart") \%>\%
      etheme_p(input$theme)
  })

  observeEvent(input$addData, {
    echartsProxy("chart") \%>\%
      eadd_linedata_p(reactive_data(), "y")
  })
}

shinyApp(ui, server)
}

}
