% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{yAxis}
\alias{eyAxis}
\alias{eyAxis_category}
\alias{eyAxis_log}
\alias{eyAxis_time}
\alias{eyAxis_value}
\alias{yAxis}
\title{Customise axis}
\usage{
eyAxis(p, show = TRUE, type = "value", append = FALSE, ...)

eyAxis_category(p, show = TRUE, zlevel = 0, z = 0, boundaryGap = FALSE,
  append = FALSE, ...)

eyAxis_value(p, show = TRUE, zlevel = 0, z = 0, position = "left",
  name = NULL, nameLocation = "end", nameTextStyle = list(),
  boundaryGap = list(0, 0), min = NULL, max = NULL, scale = FALSE,
  splitNumber = NULL, append = FALSE, ...)

eyAxis_time(p, show = TRUE, zlevel = 0, z = 0, position = "left",
  name = NULL, nameLocation = "end", nameTextStyle = list(),
  boundaryGap = list(0, 0), min = NULL, max = NULL, scale = FALSE,
  splitNumber = NULL, append = FALSE, ...)

eyAxis_log(p, show = TRUE, zlevel = 0, z = 0, position = "left",
  logLabelBase = NULL, logPositive = NULL, append = FALSE, ...)
}
\arguments{
\item{p}{an echart object.}

\item{show}{whether to show the axis.}

\item{type}{type of axis takes, \code{value}, \code{category}, \code{time}, \code{log}.}

\item{append}{whether to append options to current axis or override.}

\item{...}{any other parameter to pass to the axis.}

\item{z, zlevel}{first and second grade cascading control, the higher z the closer to the top.}

\item{boundaryGap}{whether to plot on axis line or between.}

\item{position}{position of axis, takes \code{bottom}, \code{top}, \code{left} or \code{right}.}

\item{name}{name of the axis.}

\item{nameLocation}{location of \code{name}, takes \code{start} or \code{end}.}

\item{nameTextStyle}{style \code{name} text.}

\item{min, max}{min and max values.}

\item{scale}{If \code{FALSE}, the value axis must start with 0. If \code{TRUE}, you can set the minimum and maximum value
of value axis as the starting and ending value of your value axis.}

\item{splitNumber}{number of segments, defaults to auto split along with the min/max.}

\item{logLabelBase}{base log.}

\item{logPositive}{if \code{FALSE} negative values are not supported.}
}
\description{
Customise axis
}
\examples{
df <- data.frame(x = c("Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"),
  y = runif(7, 1, 5))

df \%>\%
  echart(x) \%>\%
  eline(y) \%>\%
  exAxis_category(boundaryGap = FALSE)

df \%>\%
  echart(x) \%>\%
  ebar(y) \%>\%
  eyAxis_log()

}

